# [IMP] tests: Speed up composer tests