--- a/tests/test_helpers/dom_helper.ts
+++ b/tests/test_helpers/dom_helper.ts
@@ -140,14 +140,14 @@ export function setInputValueAndTrigger(
  */
 export async function keyDown(key: string, options: any = {}): Promise<void> {
   document.activeElement!.dispatchEvent(
-    new KeyboardEvent("keydown", Object.assign({ key, bubbles: true }, options))
+    new KeyboardEvent("keydown", Object.assign({ key, bubbles: true, cancelable: true }, options))
   );
   return await nextTick();
 }
 
 export async function keyUp(key: string, options: any = {}): Promise<void> {
   document.activeElement!.dispatchEvent(
-    new KeyboardEvent("keyup", Object.assign({ key, bubbles: true }, options))
+    new KeyboardEvent("keyup", Object.assign({ key, bubbles: true, cancelable: true }, options))
   );
   return await nextTick();
 }