… tests/components/__snapshots__/composer_integration.test.ts.snap
… tests/components/__snapshots__/top_bar.test.ts.snap
… tests/components/composer.test.ts
… tests/components/composer_integration.test.ts
… tests/plugins/edition.test.ts