## Description:

Currently, the file `composer.test.ts` takes a lot of time to execute
(43 seconds average). The reason is multi-fold:
1. There are 281 tests
2. Each test instanciate a full spreadsheet, lots of computations that
are not necessary except for ftesting the interactions of the composer
with the spreadsheet component (they currently are strongly linked).

With this revision, the tests covering the interactions with other
components are moved to a dedicated test file, which allows to test the
composer specific behaviour with a more minimalistic setup.

Note that
Benchmark
---------

`test -- composer.test.ts`
before: 43 seconds
after:  9 seconds

`test -- components`
before: 55 seconds
after: 40 seconds

Task 3163569

Odoo task ID : [3163569](https://www.odoo.com/web#id=3163569&action=333&active_id=2328&model=project.task&view_type=form&cids=1&menu_id=4720)

## review checklist

- [ ] feature is organized in plugin, or UI components
- [ ] support of duplicate sheet (deep copy)
- [ ] in model/core: ranges are Range object, and can be adapted (adaptRanges)
- [ ] in model/UI: ranges are strings (to show the user)
- [ ] undo-able commands (uses this.history.update)
- [ ] multiuser-able commands (has inverse commands and transformations where needed)
- [ ] new/updated/removed commands are documented
- [ ] exportable in excel
- [ ] translations (\_lt("qmsdf %s", abc))
- [ ] unit tested
- [ ] clean commented code
- [ ] track breaking changes
- [ ] doc is rebuild (npm run doc)
- [ ] status is correct in Odoo